requirements1.txt
torch
facenet_pytorch
pillow
flask